--- .stylelintrc.json
@@ -1,12 +1,0 @@
-{
-    "extends": "stylelint-config-standard",
-    "rules": {
-        "indentation": null,
-        "color-hex-length": null,
-        "comment-empty-line-before": null,
-        "selector-list-comma-newline-after": null,
-        "no-descending-specificity": null,
-        "selector-type-no-unknown": null,
-        "custom-property-empty-line-before": null
-    }
-}
