+++ Makefile
@@ -1,19 +1,18 @@
-all: docs/tutorial/poster.html docs/koika/poster.html docs/poster.css
+clean:
+	-rm *.css
+	-rm -fr *.{html,css}
 
-docs/poster.css: poster.less
+all: poster.css poster.html poster.pdf
+
+poster: poster.css poster.html poster.pdf
+
+poster.css: poster.less
 	lessc --strict-units=on $< $@
 
-docs/%/:
-	mkdir -p $@
+poster.html: poster.jinja2 poster.css
+	python ./test.py $< $@
 
-docs/%/poster.css: docs/poster.css | docs/%/
-	ln -f -s ../poster.css $@
-
-docs/%/poster.html: docs/%.jinja2 poster.jinja2 docs/%/poster.css | docs/%/
-	./render.py $< $@
-
-clean:
-	rm docs/poster.css
-	rm -fr docs/*/poster.{html,css}
+poster.pdf: poster.html
+	node export.js
 
 .SECONDARY:
