--- docs/koika/esh-preamble.css
@@ -1,40 +1,0 @@
-.esh-raised, .esh-raised-contents, .esh-raised-placeholder, .esh-non-ascii, .esh-non-ascii > span {
-    display: inline-block;
-    text-indent: 0;
-}
-
-.esh-non-ascii {
-    line-height: 1; /* Work around fonts whose line-height is too large, such as Latin Modern Math */
-    width: 1ch;
-}
-
-.esh-non-ascii > span {
-    left: 50%; /* Move to middle of parent */
-    position: relative;
-}
-
-.esh-non-ascii > span > span {
-    margin-left: -50%; /* Move back to center in parent */
-}
-
-.esh-raised-contents {
-    font-size: 0.9em;
-    width: 0;
-}
-
-.esh-raised-contents > span {
-    position: relative;
-}
-
-.esh-raised-placeholder {
-    height: 0;
-    visibility: hidden;
-}
-
-.esh-standalone {
-    line-height: 1;
-}
-
-.esh-strut {
-    display: inline-block;
-}
