--- docs/tutorial/poster.css
@@ -1,1 +1,0 @@
-../poster.css
