+++ poster.less
@@ -124,22 +124,22 @@ b {
         flex-direction: column;
         flex-wrap: wrap;
         overflow: auto;
-        padding: @grid-gap / 2;
+        padding: (@grid-gap / 2);
 
         // This phantom element gives a bit of breathing room when scrolling
         &::after {
             content: " ";
             display: block;
             flex-basis: 100%;
-            width: @grid-gap / 2;
+            width: (@grid-gap / 2);
         }
 
         > * {
             flex-grow: 1;
-            margin: @grid-gap / 2;
+            margin: (@grid-gap / 2);
             max-width: @column-max-width;
             min-width: @column-min-width;
-            width: 100% / @columns - 1%;
+            width: (100% / @columns - 1%);
             width: calc(100% / @columns - @grid-gap);
         }
     }
